Just Another Project Starter Template
This is a very simple project starter template, that aims to help kick-start any new design project in an organized way.

I have been using and refining this setup for a while in my day-to-day job and is working pretty well for me in many scenarios. So I hope that some remix of this could work for you as well, or at least spark some ideas to incorporate into your current approach. I tested it on small projects where you only need 1 design file, but also in mid-size ones where you probably need a different file for each section or feature of the product and having a separate file for the design library.

In this first version, you'll find:

1. File thumbnail 
2. 📄 README page: As the landing page of the project, it provides basic info along with some guidance on how to navigate this file.
3. ✅ MASTER page: To contain only the latest approved versions of the screens and flows (ready for development).
4. 🎨 Design Foundations page: To contain your live documentation and repository of all the design rules, guidelines, and reusable components.
5. [STATUS] Task page: This type of page will give you the space to explore every task in parallel, without confusing people about what's ready for development.

I would love to hear from you, so feel free to leave your comments and/or reach out via email at any time.

Thanks!